API/requirements.txt
Flask
flask-cors
firebase-admin
requests
spacy
chromadb
numpy
python-dotenv
PyMuPDF
